complete postSlice
