+++ src/store.js
@@ -1,5 +1,8 @@
 import { configureStore } from "@reduxjs/toolkit";
+import postsSlice from "./features/postsSlice";
 
 export const store = configureStore({
-  reducer: {},
+  reducer: {
+    post: postsSlice,
+  },
 });
